crawlers/requirements.txt
python-dotenv==1.0.0 
pika==1.3.2
requests==2.28.1